Fix switch on/off
If we report callback_after_update() at that moment, the device will
understand the value was already applied to the remote switch and will
request to disconnect, since its work is done.

However at that time the send_updated_params_loop() didn't have the
time to send the updates yet.